Demo codes
This is a Matlab demo codes of functional connectivity analysis based on the time-varying phase synchronization with a data-driven manner.
The basis of these programs are discribed described in the following paper (Yokoyama et al, 2018, https://www.frontiersin.org/articles/10.3389/fnhum.2018.00259/abstract).

See the details of the algrorithms in the programs for the description of the supplementary materials in above paper (correctly published
at this link: “https://www.frontiersin.org/articles/10.3389/fnhum.2018.
00259/full#supplementary-material.”).

Note that all programs are provided with Matlab format .